MasterSystem/backend/requirements.txt
fastapi
uvicorn
websockets
pydantic
loguru
requests
SpeechRecognition
edge-tts
pygame
pyaudio
numpy

google-generativeai
fer
tensorflow
python-dotenv



